refactor: FloatingLabelInput 파일 수정 후, SearchBar 컴포넌트 적용 및 버그 체크
