+++ src/app/(with-nav-footer)/activities/_components/SearchBar.tsx
@@ -1,11 +1,13 @@
 'use client';
-
 import Image from 'next/image';
 import Button from '@/components/Button';
 import bed from '@/assets/icons/bed.svg';
-import FloatingLabelInput from '@/components/FloatingLabelInput';
+import FloatingLabelInput from './FloatingLabelInput';
+import { useState } from 'react';
 
 export default function SearchBar() {
+  const [inputValue, setInputValue] = useState('');
+
   return (
     <div className='w-full min-w-[375px] px-4 md:px-6'>
       <div className='mx-auto h-[129px] w-full max-w-[1200px] min-w-[343px] justify-center rounded-2xl bg-white drop-shadow-md md:h-[184px]'>
@@ -13,7 +15,13 @@ export default function SearchBar() {
           <div className='text-black-100 text-lg font-bold md:text-xl'>무엇을 체험하고 싶으신가요?</div>
           <div className='relative mt-[15px] flex items-center gap-[12px] md:mt-[32px]'>
             <div className='relative flex-grow'>
-              <FloatingLabelInput label='내가 원하는 체험은' />
+              <FloatingLabelInput
+                id='searchId'
+                label='내가 원하는 체험은'
+                onChange={(e) => setInputValue(e.target.value)}
+                value={inputValue}
+                className='h-[56px] w-full max-w-[1004px] min-w-[187px] py-[14px] pl-[48px]'
+              />
               <Image src={bed} width={24} height={24} alt='침대' className='absolute top-[16px] left-[12px]' />
             </div>
             <Button className='px-[20px] py-[15px] whitespace-nowrap md:px-[40px] md:py-[15px]'>검색하기</Button>
