+++ src/components/Input.tsx
@@ -57,7 +57,7 @@ export type InputProps = Field &
     ref?: Ref<HTMLInputElement>;
   };
 
-export default function OauthInput({ label, error, className, ref, ...props }: InputProps) {
+export default function Input({ label, error, className, ref, ...props }: InputProps) {
   const id = useId();
 
   return (
